# Refactor parser: Avoid some assertion for StmtNode