Make code cleaner. No performance improvement.

Before:
> ➜  parser git:(master) ✗ go test -run=XXX -bench=. -benchtime=10s
BenchmarkSysbenchSelect-4   	 3000000	      4589 ns/op	    1888 B/op	      27 allocs/op
BenchmarkParse-4            	 1000000	     14310 ns/op	    5224 B/op	      89 allocs/op
PASS
ok  	github.com/pingcap/tidb/parser	32.910s
➜  parser git:(master) ✗ go test -run=XXX -bench=. -benchtime=10s
BenchmarkSysbenchSelect-4   	 3000000	      4770 ns/op	    1888 B/op	      27 allocs/op
BenchmarkParse-4            	 1000000	     14001 ns/op	    5224 B/op	      89 allocs/op
PASS
ok  	github.com/pingcap/tidb/parser	33.489s
➜  parser git:(master) ✗ go test -run=XXX -bench=. -benchtime=10s
BenchmarkSysbenchSelect-4   	 3000000	      4679 ns/op	    1888 B/op	      27 allocs/op
BenchmarkParse-4            	^[[A 1000000	     14028 ns/op	    5224 B/op	      89 allocs/op
PASS
ok  	github.com/pingcap/tidb/parser	32.791s
➜  parser git:(master) ✗ go test -run=XXX -bench=. -benchtime=10s
BenchmarkSysbenchSelect-4   	 3000000	      4628 ns/op	    1888 B/op	      27 allocs/op
BenchmarkParse-4            	 1000000	     14434 ns/op	    5224 B/op	      89 allocs/op
PASS
ok  	github.com/pingcap/tidb/parser	33.395s

After:

> ➜  parser git:(shenli/parser) ✗ go test -run=XXX -bench=. -benchtime=10s
> BenchmarkSysbenchSelect-4   	 3000000	      4586 ns/op	    1888 B/op	      27 allocs/op
> BenchmarkParse-4            	 1000000	     14229 ns/op	    5224 B/op	      89 allocs/op
> PASS
> ok  	github.com/pingcap/tidb/parser	32.819s
> ➜  parser git:(shenli/parser) ✗ go test -run=XXX -bench=. -benchtime=10s
> BenchmarkSysbenchSelect-4   	 3000000	      4567 ns/op	    1888 B/op	      27 allocs/op
> BenchmarkParse-4            	 1000000	     14253 ns/op	    5224 B/op	      89 allocs/op
> PASS
> ok  	github.com/pingcap/tidb/parser	32.768s
> ➜  parser git:(shenli/parser) ✗ go test -run=XXX -bench=. -benchtime=10s
> BenchmarkSysbenchSelect-4   	 3000000	      4622 ns/op	    1888 B/op	      27 allocs/op
> BenchmarkParse-4            	 1000000	     14979 ns/op	    5224 B/op	      89 allocs/op
> PASS
> ok  	github.com/pingcap/tidb/parser	33.719s
> ➜  parser git:(shenli/parser) ✗ go test -run=XXX -bench=. -benchtime=10s
> BenchmarkSysbenchSelect-4   	 3000000	      4701 ns/op	    1888 B/op	      27 allocs/op
> BenchmarkParse-4            	 1000000	     14687 ns/op	    5224 B/op	      89 allocs/op
> PASS
> ok  	github.com/pingcap/tidb/parser	33.645s
> ➜  parser git:(shenli/parser) ✗ go test -run=XXX -bench=. -benchtime=10s
> BenchmarkSysbenchSelect-4   	 3000000	      4561 ns/op	    1888 B/op	      27 allocs/op
> BenchmarkParse-4            	 1000000	     14233 ns/op	    5224 B/op	      89 allocs/op
> PASS
> ok  	github.com/pingcap/tidb/parser	32.730s