@shenli 2017-10-09
@XuHuaiyu I can not open CircleCI now. I merge master and rerun it.